Stop using a uuid when we need to create one
There are certain cases where we are creating a new group that doesnt
already have a ref_id. Cloudhealths api used to accept a string uuuid
but has changed to stop supporting this. Instead use an integer that is
greater than any of the integers seen when looping through the constants
section of the data structure