Add acceptance test for creation
The test creates a perspective. Terraforms test framework will take care
of destroying it - at this stage we are not actively checking it has
been destroyed - however if its not destroyed the next test run would
fail

To run the test, you must set the env var TF_ACC. It may also be useful
to see logging with TF_LOG=<level>. You must also pass the verbose flag.
e.g TF_LOG=info TF_ACC=1 go test -v

At this stage we are not really testing any properties of the created
perspective, but its useful to have this framework for adhoc testing of
what the provider is doing - you can edit the config and see what is
happening, including local code changes, which is a lot easier than the
workflow of making changes to the binary, compiling it and copying it
over the existing binary to test with some base configuration.

In order to add the acceptance tests i've had to update the vendor
directory to include the terraform helper - i've used the v0.8.8 version
of terraform for this, as this is the current version used by the
provider.

The acceptance test framework requires that after you have run a test
step, there is nothing outstanding in the plan. Because the create
method was not updating the value for constants, because this is a
computed value computed by cloudhealth, i've modified the create method
to call the read method to read back the values from cloudhealth - this
is required to make the tests pass.

Cloudhealth by default does a soft delete, and puts things into the
"trash" in the web ui when you delete, and this is the default for api
calls. Because of this i've added the hard_delete option to the schema
so the acceptance tests can do a hard delete and not leave a dangling
resource.